secureai_dataset/requirements.txt
Pillow>=9.0.0
numpy>=1.24.0
click>=8.0.0
requests>=2.28.0
opencv-python>=4.8.0
matplotlib>=3.6.0
datasets>=2.14.0
huggingface-hub>=0.16.0
